requirements.txt
# Core
streamlit
python-dotenv
numpy
pillow
opencv-python

# AI
groq

# OCR (Image reports)
paddleocr
paddlepaddle

# PDF
pdfplumber
